Moved a function to its relevant location, moved results to proper region
