+++ analysis/published_data/US/Power generation vs time.png
