+++ analysis/published_data/US/capacity_factor.csv
