+++ analysis/published_data/US/capacity_factor_extrapolated.csv
