+++ analysis/published_data/US/generation.csv
