+++ analysis/results/US/Fuel utilization.png
