+++ analysis/results/US/Number of Reactors vs Time.png
