+++ analysis/results/US/Power Generated (1971~2016).png
