+++ analysis/results/US/Power Generated.png
