+++ analysis/results/US/SWU vs Time.png
